docker/Yolo_traffic_incident/requirements.txt
opencv-python
numpy
google-cloud-pubsub
